requirements.txt
google-cloud-bigquery
google-cloud-storage
google-auth
pandas-gbq
